+++ example/lib/repository.dart
@@ -1,3 +1,5 @@
+import 'dart:math';
+
 import 'package:example/foo_bar.dart';
 import 'package:serverside_datatable/serverside_datatable.dart';
 
@@ -5,8 +7,7 @@ class Repository extends ServerSideRepository<FooBar> {
   @override
   Future<ServerSideResponse<FooBar>> fetchData(
     List<ServerSideAppliedFilter<DropdownItem>> appliedFilters,
-    FooBar? lastFetchedItem,
-    PaginateType paginateType,
+    ServerSideMetadata metadata,
     int offset,
     int limit,
   ) async {
@@ -18,6 +19,9 @@ class Repository extends ServerSideRepository<FooBar> {
     ];
     final totalRecords = fooBars.length;
     if (offset + limit > totalRecords) limit = totalRecords;
-    return ServerSideResponse(fooBars.getRange(offset, limit).toList(), totalRecords);
+    return ServerSideResponse(
+      offset > fooBars.length ? [] : fooBars.getRange(offset, limit).toList(),
+      totalRecords,
+    );
   }
 }