+++ lib/serverside_datatable.dart
@@ -1,8 +1,9 @@
 library serverside_datatable;
 
-export 'src/server_side_paginate_type.dart';
 export 'src/serverside_column.dart';
 export 'src/serverside_datatable.dart';
 export 'src/serverside_filter.dart';
 export 'src/serverside_matchmode.dart';
+export 'src/serverside_metadata.dart';
+export 'src/serverside_paginate_type.dart';
 export 'src/serverside_repository.dart';