+++ lib/src/server_side_paginate_type.dart
@@ -1,1 +1,0 @@
-enum PaginateType { next, previous }