+++ lib/src/serverside_datasource.dart
@@ -1,8 +1,9 @@
 import 'package:flutter/material.dart';
 
-import 'server_side_paginate_type.dart';
 import 'serverside_column.dart';
 import 'serverside_filter.dart';
+import 'serverside_metadata.dart';
+import 'serverside_paginate_type.dart';
 import 'serverside_repository.dart';
 
 class ServerSideDataSource<T> extends DataTableSource {
@@ -12,7 +13,7 @@ class ServerSideDataSource<T> extends DataTableSource {
   final List<ServerSideAppliedFilter> _filters = [];
   final ServerSideRepository<T> _repository;
   final Function(T rowData)? onRowClick;
-  PaginateType _paginateType = PaginateType.next;
+  ServerSidePaginateType _paginateType = ServerSidePaginateType.next;
 
   void onRepositoryParamUpdated() {
     _fetch();
@@ -21,9 +22,9 @@ class ServerSideDataSource<T> extends DataTableSource {
   set offset(offset) {
     if (_offset != offset) {
       if (offset > _offset) {
-        _paginateType = PaginateType.next;
+        _paginateType = ServerSidePaginateType.next;
       } else {
-        _paginateType = PaginateType.previous;
+        _paginateType = ServerSidePaginateType.previous;
       }
       _offset = offset;
       _fetch();
@@ -66,10 +67,10 @@ class ServerSideDataSource<T> extends DataTableSource {
   }
 
   void _fetch() async {
-    final lastFetchedItem = _data.isNotEmpty ? _data.last : null;
+    final metadata = ServerSideMetadata<T>(_paginateType, List.from(_data));
     _data.clear();
     notifyListeners();
-    _repository.fetchData(filters, lastFetchedItem, _paginateType, _offset, _rowsPerPage).then((response) {
+    _repository.fetchData(filters, metadata, _offset, _rowsPerPage).then((response) {
       _data.addAll(response.data);
       _totalRecords = response.totalRecords;
       notifyListeners();