+++ lib/src/serverside_paginate_type.dart
@@ -1,0 +1,1 @@
+enum ServerSidePaginateType { next, previous }